requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
numpy>=1.26.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
requests>=2.32.0
